// backend/2-vertical-slice/features/update-player/update-player.dto.ts
import { PlayerData } from '../../../shared/types/player.types.js';

export interface UpdatePlayerRequest {
  name?: string;
  nickname?: string;
  gender?: 'M' | 'F' | 'Misto';
  position?: PlayerData['position'];
  skillLevel?: number;
  height?: number;
  dominantHand?: 'Destro' | 'Canhoto' | 'Ambidestro';
  avatar?: string;
  teamId?: string | null;
  bio?: string;
}